refactor: names and places of things
- also allow Serialize on all response structs
