+++ src/response.rs
@@ -1,10 +1,9 @@
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
 pub mod batch;
-pub mod compatibility;
 pub mod meta;
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Access {
     pub access_token: String,
     pub expires_in: i32,
@@ -12,38 +11,38 @@ pub struct Access {
     pub token_type: String,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct ApplicationPermissions {
     pub permissions: Vec<String>,
     pub paging: Paging,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct EngineOilLife {
     pub life_remaining: f32,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct BatteryCapacity {
     pub capacity: f32,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct BatteryLevel {
     pub percent_remaining: f32,
     pub range: f32,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct ChargingStatus {
     pub is_plugged_in: bool,
     pub state: String,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct FuelTank {
     pub range: f32,
@@ -51,30 +50,30 @@ pub struct FuelTank {
     pub amount_remaining: f32,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Location {
     pub latitude: f32,
     pub longitude: f32,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Odometer {
     pub distance: f32,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Paging {
     pub count: i32,
     pub offset: i32,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Vehicles {
     pub vehicles: Vec<String>,
     pub paging: Paging,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct TirePressure {
     pub front_left: f32,
@@ -83,7 +82,7 @@ pub struct TirePressure {
     pub back_right: f32,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct VehicleAttributes {
     pub id: String,
     pub make: String,
@@ -91,18 +90,33 @@ pub struct VehicleAttributes {
     pub year: i32,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Vin {
     pub vin: String,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Action {
     pub message: String,
     pub status: String,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Disconnect {
     pub status: String,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct Capability {
+    permission: String,
+    endpoint: String,
+    capable: bool,
+    reason: Option<String>,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct Compatibility {
+    compatible: bool,
+    reason: Option<String>,
+    capabilities: Vec<Capability>,
 }
