+++ src/response/meta.rs
@@ -1,8 +1,8 @@
 use chrono::{DateTime, TimeZone, Utc};
 use reqwest::header::HeaderMap;
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
-#[derive(Deserialize, Debug)]
+#[derive(Debug, Deserialize, Serialize)]
 pub struct Meta {
     #[serde(rename = "sc-data-age")]
     pub data_age: Option<DateTime<Utc>>,
